Move controller logic to service layer
