+++ lib/galaxy/webapps/galaxy/services/base.py
@@ -6,6 +6,8 @@ from typing import (
 from galaxy.managers.base import (
     decode_with_security,
     encode_with_security,
+    get_class,
+    get_object,
     SortableManager,
 )
 from galaxy.schema.fields import EncodedDatabaseIdField
@@ -57,3 +59,15 @@ class ServiceBase:
         if order_by_query and ORDER_BY_SEP_CHAR in order_by_query:
             return [manager.parse_order_by(o) for o in order_by_query.split(ORDER_BY_SEP_CHAR)]
         return manager.parse_order_by(order_by_query)
+
+    def get_class(self, class_name):
+        """
+        Returns the class object that a string denotes. Without this method, we'd have to do eval(<class_name>).
+        """
+        return get_class(class_name)
+
+    def get_object(self, trans, id, class_name, check_ownership=False, check_accessible=False, deleted=None):
+        """
+        Convenience method to get a model object with the specified checks.
+        """
+        return get_object(trans, id, class_name, check_ownership=check_ownership, check_accessible=check_accessible, deleted=deleted)
